pkg/oscap: refactor the oscap configs
The oscap tailoring and remediation configs were too tightly coupled
with the osbuild stage implementations. These configs can be refactored
further by making the `TailoringConfig` a pointer on the
`RemediationConfig`, reversing the dependency of the configs.

These changes should reduce the complexity of the OpenSCAP stage options
and hopefully reduce some of the confusion too.